客户端控制界面给开关、温度、风速绑定通信方法

+++ flask_project1.0/db.py
@@ -56,7 +56,7 @@ class Bill(mydb.Model):
     room = mydb.Column(mydb.String(255), primary_key=True)
     start_time = mydb.Column(mydb.Integer, primary_key=True)
     end_time = mydb.Column(mydb.Integer, primary_key=True)
-    cost = mydb.Column(mydb.Double, nullable=False)
+    cost = mydb.Column(mydb.Float, nullable=False)
 
     def __repr__(self):
         return '<room %r>' % self.room
